requirements.txt
termcolor
boto3
ConfigParser
argparse
ruamel.yaml